test: Remove Vagrant VM provisioning logic
The vagrant artifacts necessary to provision the VM from scratch was
removed in commit bafb49b7379f ("cleanup: Remove vagrant related file").
Remove the -cilium.reprovision flag and corresponding logic which is now
broken and unusable.
