+++ test/k8s/net_policies.go
@@ -1398,7 +1398,7 @@ var _ = SkipDescribeIf(func() bool {
 // This Describe block is needed to run some tests in GKE. For example, the
 // kube-apiserver policy matching feature needs coverage on GKE as there are
 // two cases for that feature:
-//   - kube-apiserver running within the cluster (Vagrant VMs)
+//   - kube-apiserver running within the cluster (kind)
 //   - kube-apiserver running outside of the cluster (GKE)
 var _ = SkipDescribeIf(helpers.DoesNotRunOn54OrLaterKernel,
 	"K8sPolicyTestExtended", func() {
@@ -1516,7 +1516,7 @@ var _ = SkipDescribeIf(helpers.DoesNotRunOn54OrLaterKernel,
 						By("Checking ingress connectivity from k8s1 node to k8s1 pod (host)")
 					default:
 						// We need to bypass this check as in a non-managed
-						// environment like Vagrant, the kube-apiserver is
+						// environment like kind, the kube-apiserver is
 						// running locally on K8s1. This means that local host
 						// traffic cannot be disambiguated from kube-apiserver
 						// traffic.
